LayoutTests/imported/w3c/web-platform-tests/offscreen-canvas/text/2d.text.baseline.default-expected.txt

2d.text.baseline.default


FAIL OffscreenCanvas test: 2d.text.baseline.default assert_equals: ctx.textBaseline === 'alphabetic' (got [undefined], expected alphabetic[string]) expected (string) "alphabetic" but got (undefined) undefined
